# store/openapi.yaml
openapi: 3.1.0
info:
  title: ProAgentStore API
  description: |
    REST API for ProAgentStore — server-powered AI agents on Cloudflare Workers.
    Agents are backed by Durable Objects; each user gets a personal instance with
    isolated memory, knowledge base, and message history.

    **Auth:** All protected endpoints require a Bearer token in the `Authorization`
    header. Obtain a token via `POST /v1/auth/exchange` (FAS OAuth flow) or
    `POST /v1/auth/github` (direct GitHub OAuth).
  version: 1.0.0
  contact:
    url: https://proagentstore.online
servers:
  - url: https://api.proagentstore.online
    description: Production
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Success:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          example: true
    User:
      type: object
      properties:
        id:
          type: string
        login:
          type: string
        name:
          type: string
        avatar:
          type: string
          format: uri
        roles:
          type: array
          items:
            type: string
          example: [user, creator]
        hasSubscription:
          type: boolean
        bio:
          type: string
        website:
          type: string
        twitter:
          type: string
...
    Notification:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
        title:
          type: string
        body:
          type: string
        agent_id:
          type: string
          nullable: true
        read:
          type: integer
          enum: [0, 1]
        created_at:
          type: string
          format: date-time
...
paths:
...
  /v1/auth/exchange:
    post:
      summary: Exchange FAS token for PAGS token
      description: |
        Exchange a FreeAppStore session token for a ProAgentStore session token.
        This is the primary auth flow: complete GitHub OAuth on FAS, then call
        this endpoint with the resulting `fas_session` token.
      operationId: exchangeToken
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fas_session]
              properties:
                fas_session:
                  type: string
                  description: Bearer token from FAS OAuth
      responses:
        "200":
          description: PAGS session token issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: "#/components/schemas/User"
        "400":
          description: fas_session missing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid FAS session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/auth/github:
    post:
      summary: Direct GitHub OAuth exchange
      description: |
        Exchange a GitHub OAuth code for a PAGS token. Only works when PAGS has
        its own OAuth app configured. Prefer `/v1/auth/exchange` instead.
      operationId: githubAuth
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                  description: GitHub OAuth authorization code
                return_to:
                  type: string
                  description: URL to redirect to after auth
      responses:
        "200":
          description: PAGS session token issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: "#/components/schemas/User"
                  return_to:
                    type: string
        "401":
          description: OAuth failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "501":
          description: GitHub OAuth not configured on this server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/auth/me:
    get:
      summary: Get current user
      description: Verify the current session and return the authenticated user's profile.
      operationId: getMe
      tags: [Auth]
      responses:
        "200":
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated or token expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      summary: Update profile
      description: Update display name, bio, website, or Twitter handle.
      operationId: updateMe
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name:
                  type: string
                bio:
                  type: string
                website:
                  type: string
                twitter:
                  type: string
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "400":
          description: Nothing to update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── Agents ────────────────────────────────────────────────────────────────
...
  /v1/notifications:
    get:
      summary: List notifications
      operationId: listNotifications
      tags: [Notifications]
      parameters:
        - name: unread
          in: query
          schema:
            type: boolean
          description: If true, return only unread notifications
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
      responses:
        "200":
          description: Notification list
          content:
            application/json:
              schema:
                type: object
                properties:
                  notifications:
                    type: array
                    items:
                      $ref: "#/components/schemas/Notification"
                  unreadCount:
                    type: integer
...
  /v1/notifications/{id}/read:
    post:
      summary: Mark notification as read
      operationId: markNotificationRead
      tags: [Notifications]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"

  # ── Billing ───────────────────────────────────────────────────────────────
...